feat: Add katalog posebnih događaja module
Introduces a new module for managing a catalog of special events. This includes defining new types for events, document types, positions, and accounting entries, as well as integrating the new page into the application's navigation and state management.

The `uuid` library is added to facilitate unique ID generation for the new entities. Minor UI adjustments were made to modals to improve accessibility and user experience.