feat: Add storno field to kontiranje stavka
Introduces a 'storno' boolean field to the kontiranje stavka (accounting entry item) to indicate if an entry is reversed or voided. This change includes updates to the data type definition, the kontiranje stavka modal component for editing and displaying the new field, and the relevant table in the katalog posebnih događanja page.